Add matching over item attributes
rip AmountMatcher being simple

Overall the resulting code is simple, but it took tons of studying
the minecraft wiki and spigot javadocs to come up with a (semi-)simple
solution.

+++ src/main/java/vg/civcraft/mc/civmodcore/itemHandling/itemExpression/amount/AnyAmount.java
@@ -7,7 +7,7 @@ package vg.civcraft.mc.civmodcore.itemHandling.itemExpression.amount;
  */
 public class AnyAmount implements AmountMatcher {
 	@Override
-	public boolean matches(int amount) {
+	public boolean matches(double amount) {
 		return true;
 	}
 }
