+++ src/main/java/vg/civcraft/mc/civmodcore/itemHandling/itemExpression/amount/ExactlyAmount.java
@@ -6,14 +6,14 @@ package vg.civcraft.mc.civmodcore.itemHandling.itemExpression.amount;
  * @author Ameliorate
  */
 public class ExactlyAmount implements AmountMatcher {
-	public ExactlyAmount(int amount) {
+	public ExactlyAmount(double amount) {
 		this.amount = amount;
 	}
 
-	public int amount;
+	public double amount;
 
 	@Override
-	public boolean matches(int amount) {
+	public boolean matches(double amount) {
 		return this.amount == amount;
 	}
 }
